SVNリポジトリ内のファイルを直接開く等、ローカルプロジェクトに存在しないリソースを開いた時に、ResourceExceptionが飛ばないように対策。
git-svn-id: https://www.seasar.org/svn/chura/trunk@841 6a3759fa-7269-4e70-8ae1-ab844f716fa1

+++ dolteng/org.seasar.dolteng.eclipse.plugin/src/main/java/org/seasar/dolteng/eclipse/operation/SqlMarkingJob.java
@@ -99,6 +99,10 @@ public class SqlMarkingJob extends WorkspaceJob {
         try {
             if (unit.exists()) {
                 final IResource resource = unit.getResource();
+                if(! resource.exists()) {
+                    return Status.OK_STATUS;
+                }
+                
                 resource.deleteMarkers(Constants.ID_SQL_MAPPER, true,
                         IResource.DEPTH_ZERO);
